Updated code in enc_dec.py
